Moce Cache into separate file.
